requirements.txt
flask==3.0.0
selenium==4.15.2
google-generativeai==0.3.1
webdriver-manager==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0